hhhhh

+++ src/database/moderation.rs
@@ -30,14 +30,14 @@ impl Backend {
         Ok(result)
     }
 
-    pub(crate) async fn find_ban_entry(&self, user_id: i64) -> Result<Option<BanEntry>, Box<dyn std::error::Error>> {
+    pub(crate) async fn find_ban_entry(&self, user_id: u64) -> Result<Option<BanEntry>, Box<dyn std::error::Error>> {
         let database = self.get_database();
 
         let collection: Collection<BanEntry> = database.collection("bannedplayers");
 
         let result = collection.find_one(
             doc! { 
-                "userId": user_id
+                "userId": user_id as i64
             },
             None
         ).await?;
@@ -45,13 +45,13 @@ impl Backend {
         Ok(result)
     }
 
-    pub async fn ban_player(&self, user_id: i64, duration_in_minutes: i64, moderator: &str, reason: &str) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    pub async fn ban_player(&self, user_id: u64, duration_in_minutes: i32, moderator: &str, reason: &str) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
         let database = self.get_database();
 
         let collection: Collection<BanEntry> = database.collection("bannedplayers");
 
         let time_now: i64 = datetime_now() as i64;
-        let banned_until = if duration_in_minutes != -1 { time_now as i64 + (duration_in_minutes * 60)} else { -1 };
+        let banned_until = if duration_in_minutes != -1 { time_now as i64 + (duration_in_minutes * 60) as i64} else { -1 };
         if let Ok(Some(_)) = self.find_ban_entry(user_id).await {
             let update = doc! { "$set": doc! {
                 "bannedTime": time_now,
@@ -61,11 +61,11 @@ impl Backend {
             } };
 
             collection.update_one(doc! { 
-                "userId": user_id
+                "userId": user_id as i64
             }, update, None).await?;
         } else {
             collection.insert_one(BanEntry {
-                user_id: user_id,
+                user_id: user_id as i64,
                 banned_time: time_now,
                 banned_until: banned_until,
                 moderator: moderator.to_string(),
@@ -76,13 +76,13 @@ impl Backend {
         Ok(())
     }
 
-    pub async fn unban_player(&self, user_id: i64) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    pub async fn unban_player(&self, user_id: u64) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
         let database = self.get_database();
 
         let collection: Collection<BanEntry> = database.collection("bannedplayers");
 
         if let Ok(Some(_)) = self.find_ban_entry(user_id).await {
-            collection.delete_one(doc! { "userId": user_id }, None).await?;
+            collection.delete_one(doc! { "userId": user_id as i64}, None).await?;
         }
 
         Ok(())
